requirements.txt
PyYAML==6.0.1
watchdog==4.0.0
pytest==8.0.0